moved TestCases into Namespaces
