+++ Assets/Scripts/Extentions/VectorToSetupPointExtention.cs
@@ -1,8 +1,7 @@
-﻿
-using JsonObjects;
+﻿using TestScenarios.JsonClasses;
 using UnityEngine;
 
-namespace DefaultNamespace
+namespace Extentions
 {
     public static class VectorToSetupPointExtention
     {
@@ -13,6 +12,18 @@ namespace DefaultNamespace
                 x = point.x, 
                 y = point.y, 
                 z = point.z
+            };
+        }
+        
+        public static SetupPoint ToVector(this Quaternion quaternion)
+        {
+            var angles = quaternion.eulerAngles;
+
+            return new SetupPoint()
+            {
+                x = angles.x,
+                y = angles.y, 
+                z = angles.z
             };
         }
     }
