--- Assets/Scripts/TestController/Tests/example.json
@@ -1,128 +1,0 @@
-﻿{
-  "scenarios": [
-    {
-      "time" : 12.0,
-          "name" : "Neuer_Test",
-          "startValues": [
-            {
-              "gameObject": "car",
-              "position": {
-                "x": 0.0,
-                "y": 0.0,
-                "z": 0.0
-              },
-              "velocity": {
-                "x": 0.0,
-                "y": 0.0,
-                "z": 0.0
-              },
-              "angularVelocity": {
-                "x": 0.0,
-                "y": 0.0,
-                "z": 0.0
-              },
-          "rotation": {
-            "x": 0.0,
-            "y": 0.0,
-            "z": 0.0
-          }
-        },
-        {
-          "gameObject": "ball",
-          "position": {
-            "x": 5.0,
-            "y": 0.0,
-            "z": 0.0
-          },
-          "velocity": {
-            "x": 0.0,
-            "y": 0.0,
-            "z": 0.0
-          },
-          "angularVelocity": {
-            "x": 0.0,
-            "y": 0.0,
-            "z": 0.0
-          },
-          "rotation": {
-            "x": 0.0,
-            "y": 0.0,
-            "z": 0.0
-          }
-        }
-      ],
-      "actions": [
-        {
-          "duration": 0.1,
-          "inputs": [
-            {
-              "name": "jump",
-              "value": 1.0
-            }
-          ]
-        },
-        {
-          "duration": 0.9,
-          "inputs": [
-          ]
-        },
-        {
-          "duration": 1.0,
-          "inputs": [
-            {
-              "name": "boost",
-              "value": 1.0
-            },
-            {
-              "name": "steer",
-              "value": 1.0
-            }
-          ]
-        },
-        {
-          "duration": 1.0,
-          "inputs": [
-            {
-              "name": "throttle",
-              "value": -1.0
-            },
-            {
-              "name": "steer",
-              "value": -1.0
-            }
-          ]
-        },
-        {
-          "duration": 3.0,
-          "inputs": [
-            {
-              "name": "throttle",
-              "value": 1.0
-            },
-            {
-              "name": "handbrake",
-              "value": 1.0
-            },
-            {
-              "name": "steer",
-              "value": 1.0
-            }
-          ]
-        },
-        {
-          "duration": 1.0,
-          "inputs": [
-            {
-              "name": "jump",
-              "value": 1.0
-            },
-            {
-              "name": "roll",
-              "value" : -1.0
-            }
-          ]
-        }
-      ]
-    }
-  ]
-}
